rename some sq::Material methods Pipeline now stores a pointer to the PassConfig it was created with Pipeline colourBlend option is now an array, one entry for each colour attachment add stencil information to PassConfig Texture filter/mipmaps modes are now enums Texture::load_from_memory now takes the mipmap level as a parameter allow loading Radiance HDR images into E5B9G9R9 textures add support for loading texture mipmaps from files  - currently only implemented for cubemaps fix mipmaps not actually being used at all due to wrong sampler setup

+++ include/sqee/vk/Helpers.hpp
@@ -122,7 +122,7 @@ inline void vk_update_descriptor_set (
 ) {
     ctx.device.updateDescriptorSets (
         vk::WriteDescriptorSet {
-            dset, binding, index, bufferInfo.size(), type, nullptr, bufferInfo.data(), nullptr
+            dset, binding, index, uint(bufferInfo.size()), type, nullptr, bufferInfo.data(), nullptr
         }, {}
     );
 }
@@ -133,10 +133,10 @@ inline void vk_update_descriptor_set_swapper (
     ctx.device.updateDescriptorSets (
         {
             vk::WriteDescriptorSet {
-                swapper.front, binding, index, bufferInfoFront.size(), type, nullptr, bufferInfoFront.data(), nullptr
+                swapper.front, binding, index, uint(bufferInfoFront.size()), type, nullptr, bufferInfoFront.data(), nullptr
             },
             vk::WriteDescriptorSet {
-                swapper.back, binding, index, bufferInfoBack.size(), type, nullptr, bufferInfoBack.data(), nullptr
+                swapper.back, binding, index, uint(bufferInfoBack.size()), type, nullptr, bufferInfoBack.data(), nullptr
             },
         }, {}
     );
